refactor some more testing
