+++ src/app/features/tracker/components/history-entry/history-entry.component.spec.ts
@@ -21,7 +21,7 @@ import {
     UpdateHistoryItemTimesPayload
 } from '../../../../shared/models';
 
-import { user } from '../../../../testing';
+import { user } from '../../../../test-helpers';
 
 describe('HistoryEntryComponent', () => {
   let store: Store<fromRoot.State>;
