+++ src/app/shared/utils/history-filter.utils.spec.ts
@@ -5,7 +5,7 @@ import {
     getUniqueFrom
 } from './history-filter.utils';
 
-import { history } from '../../testing';
+import { history } from '../../test-helpers';
 
 describe('History Filter Utils', () => {
   describe('getUniqueFrom', () => {
