+++ src/app/test-helpers/index.ts
@@ -2,3 +2,4 @@ export * from './user';
 export * from './auth';
 export * from './tracker';
 export * from './history';
+export * from './routing';
