+++ src/app/test-helpers/routing.ts
@@ -1,0 +1,13 @@
+export namespace routing {
+  export class MockRouterStateSnapshot {
+    url = 'app';
+  }
+
+  export const mockActivatedRoute = {
+    snapshot: {
+      queryParams: {
+        returnUrl: ''
+      }
+    }
+  };
+}
